src/main/java/gbsio/esiclient/client/request/industry/MiningLedgerRequest.java
package gbsio.esiclient.client.request.industry;

import com.fasterxml.jackson.core.type.TypeReference;
import com.google.common.base.Preconditions;
import com.google.common.collect.ImmutableList;
import gbsio.esiclient.api.domain.response.industry.MiningReport;
import gbsio.esiclient.api.request.specific.GetListRequest;

import java.util.Optional;

/**
 * Retrieves a character's personal mining ledger.
 */
final public class MiningLedgerRequest implements GetListRequest<MiningReport, ImmutableList<MiningReport>> {
    private final int characterID;
    private final String accessToken;

    /**
     * @param characterID the ID of the character to query
     * @param accessToken an OAuth2 access token belonging to said character
     */
    public MiningLedgerRequest(final int characterID, final String accessToken) {
        this.characterID = characterID;
        this.accessToken = Preconditions.checkNotNull(accessToken);
    }

    @Override
    public String getURL() {
        return String.format("/v1/characters/%d/mining/", characterID);
    }

    @Override
    public TypeReference<ImmutableList<MiningReport>> getExpectedType() {
        return new TypeReference<ImmutableList<MiningReport>>() {
        };
    }

    @Override
    public Optional<String> getAccessToken() {
        return Optional.of(accessToken);
    }
}
